refactor: reorganize tests for patiente templates endpoints
